Add platform information to Travis::Build::Data
This is work in progress.

It is debatable whether this is the right place to put this
information. For now, let us assume that for which platform we
should construct the build script will be included in the payload.

Travis::Build::Script will delegate the associated calls to
Travis::Build::Data.